fixed initialization of algorithm and computation of the max flow

+++ include/LAGraphX.h
@@ -1498,7 +1498,7 @@ int LAGraph_MaxFlow(
     GrB_Index S, //source node index
     GrB_Index T, // sink node index
     //outputs
-    int* f,
+    float* f,
     //inout
     char* msg
 );
